+++ src/theme/index.ts
@@ -4,12 +4,14 @@ import global from "./foundations/global";
 import typography from "./foundations/typography";
 import { spacing } from "./foundations/spacing";
 import breakpoints from "./foundations/breakpoints";
+import Accordion from "./components/accordion";
 import Button from "./components/button";
 import Checkbox from "./components/checkbox";
 import Radio from "./components/radio";
 import CustomRadioGroup from "./components/customRadioGroup";
 import CustomCheckboxGroup from "./components/customCheckboxGroup";
 import Tabs from "./components/tabs";
+import CustomSelect from "./components/select";
 import Heading from "./components/heading";
 
 /**
@@ -41,12 +43,14 @@ const theme = extendTheme({
    * https://chakra-ui.com/docs/theming/component-style
    */
   components: {
+    Accordion,
     Button,
     Checkbox,
     CustomCheckboxGroup,
     Heading,
     Radio,
     CustomRadioGroup,
+    CustomSelect,
     Tabs,
   },
   // Chakra prefixes its own CSS variables with `--chakra` by default but this